Part4
Consider the case:
Enter the word: Father
['A male parent.', "A person that has founded or originated, as in ''the father of our country''", 'The head of an organized crime fa
mily.']
Enter the word: Fatherl
Check the word again. It doesn't exist.
Enter the word: Fatherer
Check the word again. It doesn't exist.

This might be a typo or spelling mistake if a person is not so familiar with the English Language.

For that, we can calculate the similarity ratio and give the best suggestion to the user.