Part6
The progam is made more user-friendly like its showing definitions in different lines, not as a list.

Also, I have taken care of proper noun such as places, name etc. If  a user has to check for 'London', the system won't check for 'london' provided the user will enter the first letter of the proper noun always in capital letters.

Using .isupper function take care of acronyms as well.

This snippet of code can be attached to database for faster and economic execution instead of a json file. Its a CLI app for now, we can move on to GUI or Web app with some minor modifications.